Only set the root logger’s level if it has none set yet.
This needs to be done to prevent overriding the user’s configuration which
can e.g. be done by using a properties file. If the user set the log level
of our own root logger to e.g. INFO this would simply thrash that
configuration and reset the level back to ALL. Do not want.