Add ToolTipList to Swing utilities.
The ToolTipList can delegate the retrieval of the tooltip text to the component
returned by the cell renderer of the list. If the component used to render the list
is a Container the tool tip of the component under the mouse cursor is asked for
its tooltip which is then displayed for the list.